# docs/swagger.yaml
basePath: /
definitions:
...
  models.Area:
    properties:
      area_coverage:
        type: number
      area_name:
        type: string
      branch:
        $ref: '#/definitions/models.Branch'
      branch_id:
        type: integer
      created_by:
        type: string
      created_on:
        type: string
      district_coverage:
        type: number
      district_id:
        type: string
      id:
        description: swagger:model Area
        type: integer
      updated_by:
        type: string
      updated_on:
        type: string
    type: object
  models.Branch:
    properties:
      aashram_area:
        type: number
      address:
        type: string
      city:
        type: string
      contact_number:
        type: string
      coordinator_name:
        type: string
      country:
        type: string
      created_by:
        type: string
      created_on:
        type: string
      daily_end_time:
        type: string
      daily_start_time:
        type: string
      district:
        type: string
      email:
        type: string
      established_on:
        type: string
      id:
        type: integer
      name:
        type: string
      open_days:
        description: e.g. "Mon-Fri"
        type: string
      pincode:
        type: string
      police_station:
        type: string
      post_office:
        type: string
      state:
        type: string
      updated_by:
        type: string
      updated_on:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: vaish@gmail.com
    name: API Support
  description: API documentation for the DJJS Event Reporting Backend.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: DJJS Event Reporting API
  version: "1.0"
paths:
  /api/areas:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Area'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all areas
      tags:
      - Areas
    post:
      consumes:
      - application/json
      parameters:
      - description: Area payload
        in: body
        name: area
        required: true
        schema:
          $ref: '#/definitions/models.Area'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new area
      tags:
      - Areas
  /api/areas/{id}:
    delete:
      parameters:
      - description: Area ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete an area
      tags:
      - Areas
    put:
      consumes:
      - application/json
      parameters:
      - description: Area ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated fields
        in: body
        name: area
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update an area
      tags:
      - Areas
  /api/areas/search:
    get:
      description: Retrieve an area by its name, or all areas if no name is provided.
      parameters:
      - description: Area Name
        in: query
        name: area_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Area'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get areas by name (or all if none provided)
      tags:
      - Areas
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"